New API: experimental.Cows (#23)
Added Cows class in the new submodule experimental. It can compute
classic sWeights and COWs. Summation is used whenever possible and safe.
Always does the best thing possible based on what the user provided as
input for the calculation. The optimal normalization function is
computed automatically based on the provided signal and background pdfs
and a provided sample (via an extended unbinned NLL fit). Automatic
choices can be overridden with special keywords.

After another round of edits it is now possible to pass parametric pdfs
to the Cows class, which are then fitted (not only the amplitudes but
also the shapes) to the provided sample. Whether the fit is successful
depends on the bounds and starting values that the user has to provide.
The code cannot guess bounds, but starting values are based on an
educated guess if bounds are provided, if the user does not explicitly
specify the starting values.