Fixed a bug when checking package version with non-standard character (#27)
When converting a RooFit pdf to a python wrapper using convert_rf_pdf a
check is performed on the ROOT version.
If the ROOT version contains a non-standard character, eg. `/`, then the
check will produce an error. For example using ROOT version 6.30/04
reproduces this issue.

This fix converts `/` to `.` in the the version string so that is is parsed correctly. A test is added to confirm that `import_optional_module` now works correctly.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Hans Dembinski <hans.dembinski@gmail.com>